Moved `EventLoop` to `events::event_loop.rs`
