+++ wolf_engine_core/src/events/event_loop.rs
@@ -1,0 +1,5 @@
+pub trait EventLoop<E> {
+    fn next_event(&self) -> Option<E>;
+    fn send_event(&self, event: E);
+}
+
