+++ wolf_engine_core/src/events/event_queue.rs
@@ -1,6 +1,6 @@
 use std::sync::mpsc::{channel, Receiver, Sender};
 
-use crate::EventLoop;
+use crate::events::EventLoop;
 
 /// Provides a generic, fifo, mpsc event queue based on [std::sync::mpsc].
 ///
