+++ wolf_engine_core/src/events/mod.rs
@@ -2,4 +2,6 @@
 
 mod event_queue;
 pub use event_queue::*;
+mod event_loop;
+pub use event_loop::*;
